[XPU][TritonIntelGPUToLLVM] Handle special "unbroadcast" case
Unbroadcast is a special case when converting from a slice layout to a non-sliced layout.
In particular, when the required transformation simply implies "dropping" some values
owned by a given thread as they are already handled by other threads in the same warp.

As an example, going from a sliced layout with a single warp in which all elements
are owned by all threads to a blocked layout in which elements are distributed across threads
is an unbroadcast conversion.

Implement this as a sequence of select operations using the sub-group local ID as a discriminator.
This actually leads to good codegen in IGC.

Additionally, modify analyses so no SLM is allocated for this kind of conversions and thus no
barrier is introduced either.

Signed-off-by: victor-eds <victor.perez@codeplay.com>